Major update
Implemented Timewise test-train split

Updated FaultClustering to be less computational intensive.
Moved FaultClustering to K-CV for quantitywise split
Made it only use training data as input for timewise split.
FaultClustering is now a MUST HAVE in  model, if removed from 'featuresToKeep', it stops working, but this can somehow easily be changed.

Implemented plot that shows high fault zones on the network

Implemented Pressure V1.0  as a linear correlation. However this needs to be fixed in future version, as the pressure is not accurate for the measured values at one of the measurement stations.
Functions for Topography has been made, but is not implemented in the code.

Implemented code to investigate Hofors joint data.

Updated the model to automatically change the 'bestModel' estimator to be the one with the best results in K-CV.