FaultClustering in 'HPTuning' and automatic cut-off threshhold
Updated the final plots in 'RunModel' to automatically update the cut-off threshold depending on the value that grants the highest evaluation value. The standard value currently used is the AUC.
Additionally the plot of the likelyhood of pipe leakages for each pipe type is upadted with the cut-off threshhold displayed.

Updated 'HPTuning' in 'model.py' to correctly do fault clustering of both validation and test data for both quantity split and timewise split.